pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.59.0",
    "arxiv>=2.2.0",
    "mcp>=1.12.2",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
]
